Switched dataset marker resolution to a lock-guarded runtime
-- Removed dataset marker tracker file and dataset marker resolution pipeline.
-- Adjusted the dataset marker resolution runtime to acquire a .lock file and run under an exclusive lock.

Note: This and further changes to dataset marker resolution are designed to bring this runtime in-line with the project manifest generation runtime. Primarily, this is because the outcome of the dataset marker resolution pipeline can be verified by the presence or absence of the binary marker file, similar to the manifest file pipeline. This is not the case for most data processing pipelines.